pyproject.toml
[tool.poetry]
name = "aiogram-arduino-room-climat"
version = "0.1.2"
description = ""
authors = ["Yaroslav <103318612+YaSoRoP@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
environs = "^10.3.0"
pyserial = "^3.5"


[tool.poetry.scripts]
com-get-info = 'aiogram_arduino_room_climat.telegram_bot.services.poetry_function:com_get_info'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
